rename file from camel to dash
