+++ src/managers/vote-manager.ts
